updated model/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
matplotlib>=3.3.0
Pillow>=8.0.0
opencv-python>=4.4.0
scikit-image>=0.17.0
tqdm>=4.50.0

# Optional dependencies for model optimization and deployment
# onnx>=1.7.0
# onnxruntime>=1.4.0
# onnxoptimizer>=0.2.6 